pyproject.toml
[project]
name = "network-mcp"
version = "0.1.1"
description = "Network MCP Server for AI Agents - Connectivity testing and pcap analysis"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
authors = [
    { name = "Du'An Lightfoot" }
]
keywords = [
    "mcp",
    "network",
    "networking",
    "netops",
    "pcap",
    "scapy",
    "ping",
    "traceroute",
    "dns",
    "mtr",
    "observability",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Environment :: Console",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Networking",
    "Topic :: System :: Networking :: Monitoring",
]

dependencies = [
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
    "scapy>=2.5.0",
    "dnspython>=2.4.0",
    "pyyaml>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/labeveryday/network-mcp"
Repository = "https://github.com/labeveryday/network-mcp"
Issues = "https://github.com/labeveryday/network-mcp/issues"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[project.scripts]
network-mcp = "network_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/network_mcp"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-m 'not integration'"
markers = [
    "integration: requires network access and/or platform system tools",
]
